support stack objects (alloca) on stack cache
A simple escape analysis based on the PtrUseVisitor ensures that an object is
only used within the current function.  FIs are sorted so that alloca objects
are allocated at the end of the frame (higher offsets).
Instructions referring to stack objects, which would be missed by
eliminateFrameIndex() later, are rewritten to be stack cache typed right away.